WriteOutput* functions return size_t value

+++ src/host/_output.h
@@ -42,14 +42,14 @@ NTSTATUS FillOutput(SCREEN_INFORMATION& screenInfo,
 
 void FillRectangle(const CHAR_INFO * const pciFill, SCREEN_INFORMATION& screenInfo, const SMALL_RECT * const psrTarget);
 
-ULONG WriteOutputAttributes(SCREEN_INFORMATION& screenInfo,
-                            const std::vector<WORD>& attrs,
-                            const COORD target);
+size_t WriteOutputAttributes(SCREEN_INFORMATION& screenInfo,
+                             const std::vector<WORD>& attrs,
+                             const COORD target);
 
-ULONG WriteOutputStringW(SCREEN_INFORMATION& screenInfo,
-                         const std::vector<wchar_t> chars,
-                         const COORD target);
+size_t WriteOutputStringW(SCREEN_INFORMATION& screenInfo,
+                          const std::vector<wchar_t> chars,
+                          const COORD target);
 
-ULONG WriteOutputStringA(SCREEN_INFORMATION& screenInfo,
-                         const std::vector<char> chars,
-                         const COORD target);
+size_t WriteOutputStringA(SCREEN_INFORMATION& screenInfo,
+                          const std::vector<char> chars,
+                          const COORD target);
